app/backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
psycopg[binary]>=3.1
pgvector>=0.2.5
sqlalchemy>=2.0
beautifulsoup4>=4.12
lxml>=5.1
httpx>=0.27
tenacity>=8.2
rank-bm25>=0.2.2
numpy>=1.26
pyyaml>=6.0
anthropic>=0.25
sentence-transformers>=2.6
mcp>=1.0
rich>=13.7
typer>=0.12
yfinance>=0.2.36
feedparser>=6.0
scikit-learn>=1.4
